Добавлен метод showLevel. Вместо рисования всего уровня на фреймбуфере уровень перерисовывается каждый кадр. Перерисовываются только те участки, которые попадают в экран.

+++ src/animationproject/Level.java
@@ -20,5 +20,5 @@ public class Level {
     public int id;
     
     @XStreamAlias("grounds")
-    public int[][] grounds;
+    public short[][] grounds;
 }
